Speed Up Discord Bot Function
[Problem]
The Discord bot function is still a bit slow at times, resulting in
timeouts on Discord.

[Solution]
Resolve both the Discord verification and the Commands Lambda call at
the same time when needed. This saves approximately 500ms since they can
be done in parallel.

[Testing]
Deployed and manually verified that this properly resolves the issues.